spake2 code reorganization
